+++ 04-java-oop-basic/task02/src/com/example/task02/TimeSpan.java
@@ -1,0 +1,66 @@
+package com.example.task02;
+
+public class TimeSpan {
+
+    private int hours;
+    private int minutes;
+    private int seconds;
+
+    public int getHours() {
+        return hours;
+    }
+
+    public void setHours(int hours) {
+        this.hours = hours;
+    }
+
+    public int getMinutes() {
+        return minutes;
+    }
+
+    public void setMinutes(int minutes) {
+        this.minutes = minutes;
+    }
+
+    public int getSeconds() {
+        return seconds;
+    }
+
+    public void setSeconds(int seconds) {
+        this.seconds = seconds;
+    }
+
+    public TimeSpan(int hours, int minutes, int seconds) {
+
+        this.hours = hours;
+        this.minutes = minutes;
+        this.seconds = seconds;
+    }
+
+    void add(TimeSpan time) {
+
+        this.hours += time.hours;
+        this.minutes += time.minutes;
+        this.seconds += time.seconds;
+        normalizeTime();
+    }
+
+    void subtract(TimeSpan time) {
+
+        this.hours -= time.hours;
+        this.minutes -= time.minutes;
+        this.seconds -= time.seconds;
+        normalizeTime();
+    }
+
+    void normalizeTime() {
+        minutes += seconds / 60;
+        seconds = seconds % 60;
+        hours += minutes / 60;
+        minutes = minutes % 60;
+    }
+
+    public String toString() {
+        return String.format("%d,%d,%d", seconds, hours, minutes);
+    }
+}
